requirements.txt
aiofiles==23.2.1
aiogram==3.12.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
frozenlist==1.4.1
greenlet==3.0.3
idna==3.8
isodate==0.6.1
lxml==5.3.0
magic-filter==1.0.12
Mako==1.3.5
MarkupSafe==2.1.5
more-itertools==10.5.0
multidict==6.0.5
passlib==1.7.4
platformdirs==4.3.2
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
simple-salesforce==1.12.6
six==1.16.0
SQLAlchemy==2.0.34
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.7
zeep==4.2.1
